backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
uuid>=1.30
psycopg2-binary>=2.9.6